+++ src/codemagic/tools/app_store_connect.py
@@ -368,9 +368,12 @@ class AppStoreConnect(cli.CliApp):
             profile_type=profile_type,
             bundle_id=bundle_id_resource_id,
             certificates=certificate_resource_ids,
-            devices=device_resource_ids,
-            omit_keys=['devices']
         )
+        if profile_type.devices_allowed():
+            create_params.update({
+                'devices': device_resource_ids,
+                'omit_keys': ['devices']
+            })
         profile = self._create_resource(self.api_client.profiles, should_print, **create_params)
 
         if save: